backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.20
pyodbc==4.0.35
python-dotenv==1.0.0
bcrypt==4.0.1
azure-storage-blob==12.16.0
requests==2.31.0
